Refactoring code to use routers.
