+++ app/main.py
@@ -1,89 +1,17 @@
-from typing import List
+from fastapi import FastAPI
 
-from fastapi import FastAPI, status, HTTPException, Response, Depends
-from sqlalchemy.orm import Session
+from . import models
+from .database import engine
+from .routers import post, user
 
-from . import models, schemas
-from .database import engine, get_db
-from .utils import hash_password
 
 models.Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
+app.include_router(post.router)
+app.include_router(user.router)
 
 
 @app.get("/")
 def root():
     return {"message": "Hello World"}
-
-
-@app.get("/posts", response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
-    posts = db.query(models.Post).all()
-    return posts
-
-
-@app.post("/posts", status_code=status.HTTP_201_CREATED,
-          response_model=schemas.Post)
-def create_post(post: schemas.PostCreate, db: Session = Depends(get_db)):
-    new_post = models.Post(**post.dict())
-    db.add(new_post)
-    db.commit()
-    db.refresh(new_post)
-    return new_post
-
-
-@app.get("/posts/{id}", response_model=schemas.Post)
-def get_post(id: int, db: Session = Depends(get_db)):
-    post = db.query(models.Post).filter(models.Post.id == id).first()
-    if not post:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Post with id {id} was not found.")
-    return post
-
-
-@app.put("/posts/{id}", response_model=schemas.Post)
-def update_post(id: int, post: schemas.PostCreate,
-                db: Session = Depends(get_db)):
-    post_query = db.query(models.Post).filter(models.Post.id == id)
-    update_post = post_query.first()
-    if update_post is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Post with id {id} was not found.")
-    post_query.update(post.dict(), synchronize_session=False)
-    db.commit()
-    return post_query.first()
-
-
-@app.delete("/posts/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id: int, db: Session = Depends(get_db)):
-    post = db.query(models.Post).filter(models.Post.id == id)
-    if post.first() is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Post with id {id} was not found.")
-    post.delete(synchronize_session=False)
-    db.commit()
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
-
-
-@app.post("/users", status_code=status.HTTP_201_CREATED,
-          response_model=schemas.UserOut)
-def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
-
-    user.password = hash_password(user.password)
-
-    new_user = models.User(**user.dict())
-    db.add(new_user)
-    db.commit()
-    db.refresh(new_user)
-    return new_user
-
-
-@app.get('/users/{id}', response_model=schemas.UserOut)
-def get_user(id: int, db: Session = Depends(get_db)):
-    user = db.query(models.User).filter(models.User.id == id).first()
-
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"User with id {id} does not exist!")
-    return user
