+++ app/routers/user.py
@@ -1,0 +1,31 @@
+from fastapi import status, HTTPException, Depends, APIRouter
+from sqlalchemy.orm import Session
+
+from .. import models, schemas
+from ..database import get_db
+from ..utils import hash_password
+
+router = APIRouter()
+
+
+@router.post("/users", status_code=status.HTTP_201_CREATED,
+          response_model=schemas.UserOut)
+def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
+
+    user.password = hash_password(user.password)
+
+    new_user = models.User(**user.dict())
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+    return new_user
+
+
+@router.get('/users/{id}', response_model=schemas.UserOut)
+def get_user(id: int, db: Session = Depends(get_db)):
+    user = db.query(models.User).filter(models.User.id == id).first()
+
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} does not exist!")
+    return user
